Create README.md
Passionate and detail-oriented UI/UX designer with a strong foundation in user-centered
design principles and proficiency in design tools like Adobe XD and Figma. Eager to
contribute to a dynamic team to create initiative and engaging user experiences. Seeking to
join in good position to leverage my skills in wireframing, prototyping, and user research to
enhance user satisfaction and drive business growth.